Remove duplicated code in the TestableSocket.
