+++ SourceQuery/Socket/SourceSocket.php
@@ -15,35 +15,14 @@ declare(strict_types=1);
 
 namespace xPaw\SourceQuery\Socket;
 
-use xPaw\SourceQuery\Buffer;
-use xPaw\SourceQuery\Exception\InvalidPacketException;
+use xPaw\SourceQuery\Socket\Traits\SourcePacketDataTrait;
 
 final class SourceSocket extends AbstractSocket
 {
+    use SourcePacketDataTrait;
+
     public function getType(): int
     {
         return SocketType::SOURCE;
-    }
-
-    /**
-     * @throws InvalidPacketException
-     */
-    protected function readInternalPacketData(
-        Buffer $buffer,
-        int &$count,
-        int &$number,
-        bool &$isCompressed,
-        ?int &$checksum
-    ): void {
-        $count = $buffer->getByte();
-        $number = $buffer->getByte() + 1;
-
-        if ($isCompressed) {
-            $buffer->getLong(); // Split size.
-
-            $checksum = $buffer->getUnsignedLong();
-        } else {
-            $buffer->getShort(); // Split size.
-        }
     }
 }
